refactor(compaction): name the usage capability for its contract
The naming pass in #480 (issue #392) settled on an imperative verb phrase when a
capability acts on the run and one verb phrase states its whole contract, and on
avoiding agent-nouns for single-behavior capabilities. This capability reports
context usage before each model request, so `ReportContextUsage` says what it
does; the `ContextUsage` payload stays a noun. No deprecated alias: the old name
was never released.
