+++ pydantic_ai_harness/compaction/__init__.py
@@ -7,13 +7,13 @@ Each capability lives in its own module; shared utilities (token estimation, the
 from pydantic_ai_harness._warn import warn_class_renamed
 from pydantic_ai_harness.compaction._clamp_oversized_messages import ClampOversizedMessages
 from pydantic_ai_harness.compaction._clear_tool_results import ClearToolResults
-from pydantic_ai_harness.compaction._context_usage import ContextUsage, ContextUsageMonitor
 from pydantic_ai_harness.compaction._context_window import (
     DEFAULT_CONTEXT_WINDOW,
     resolve_context_window,
 )
 from pydantic_ai_harness.compaction._deduplicate_file_reads import DeduplicateFileReads
 from pydantic_ai_harness.compaction._manual import compact_now
+from pydantic_ai_harness.compaction._report_context_usage import ContextUsage, ReportContextUsage
 from pydantic_ai_harness.compaction._shared import CompactionStrategy, SupportsFocus, estimate_token_count
 from pydantic_ai_harness.compaction._sliding_window_compaction import SlidingWindowCompaction
 from pydantic_ai_harness.compaction._summarizing_compaction import SummarizingCompaction
@@ -26,7 +26,7 @@ __all__ = [
     'ClearToolResults',
     'CompactionStrategy',
     'ContextUsage',
-    'ContextUsageMonitor',
+    'ReportContextUsage',
     'DeduplicateFileReads',
     'SlidingWindowCompaction',
     'SummarizingCompaction',
